unify the fragment class names
