+++ app/src/main/java/es/wolfi/app/passman/adapters/CustomFieldViewAdapter.java
@@ -32,7 +32,7 @@ import org.json.JSONException;
 import java.util.List;
 
 import es.wolfi.app.passman.CopyTextItem;
-import es.wolfi.app.passman.fragments.CredentialDisplay;
+import es.wolfi.app.passman.fragments.CredentialDisplayFragment;
 import es.wolfi.app.passman.R;
 import es.wolfi.passman.API.CustomField;
 import es.wolfi.passman.API.File;
@@ -40,15 +40,15 @@ import es.wolfi.utils.FileUtils;
 
 /**
  * {@link RecyclerView.Adapter} that can display a {@link File} and makes a call to the
- * specified {@link CredentialDisplay.OnListFragmentInteractionListener}.
+ * specified {@link CredentialDisplayFragment.OnListFragmentInteractionListener}.
  * TODO: Replace the implementation with code for your data type.
  */
 public class CustomFieldViewAdapter extends RecyclerView.Adapter<CustomFieldViewAdapter.ViewHolder> {
 
     private final List<CustomField> mValues;
-    private final CredentialDisplay.OnListFragmentInteractionListener customFieldListListener;
+    private final CredentialDisplayFragment.OnListFragmentInteractionListener customFieldListListener;
 
-    public CustomFieldViewAdapter(List<CustomField> customFields, CredentialDisplay.OnListFragmentInteractionListener listener) {
+    public CustomFieldViewAdapter(List<CustomField> customFields, CredentialDisplayFragment.OnListFragmentInteractionListener listener) {
         mValues = customFields;
         customFieldListListener = listener;
     }
